+++ server/http.go
@@ -78,7 +78,7 @@ func New(cmdCtx, evalCtx context.Context, log logrus.FieldLogger, settings *conf
 	var serverTLS *config.ServerTLS
 	for host, muxOpts := range hosts {
 		mux := NewMux(muxOpts)
-		mux.registerHandler(mux.endpointRoot, []string{http.MethodGet}, settings.HealthPath, handler.NewHealthCheck(settings.HealthPath, shutdownCh))
+		registerHandler(mux.endpointRoot, []string{http.MethodGet}, settings.HealthPath, handler.NewHealthCheck(settings.HealthPath, shutdownCh))
 		muxersList[host] = mux
 
 		// TODO: refactor (hosts,muxOpts, etc) format type and usage